Change type of additional points from Integer to Float

+++ src/main/java/com/mafia/statistics/MafiaStatisticsAPI/dto/host/BestPlayer.java
@@ -15,5 +15,5 @@ public class BestPlayer {
 
     private Player player;
 
-    private Integer additionalPoints;
+    private Float additionalPoints;
 }
